Move aggregation to grid endpoint folder
Prevent duplicate grid requests on startup
